working dtypes guard all the way through
