+++ nvtabular/ops/hash_bucket.py
@@ -123,4 +123,4 @@ class HashBucket(Operator):
 
     @property
     def output_dtype(self):
-        return numpy.uint32
+        return numpy.int32
